requirement.txt
Django==2.14
Markdown==3.11
PyMysql==0.93
django-mdeditor==0.1.14